Make forgot-password a route instead of a client toggle
Clicking "Forgot your password?" now navigates to /forgot-password
rather than swapping the form in place. The URL reflects the state —
bookmarkable, back button works — and each page's server component
owns its own description, so no mode-aware copy is needed.

Removes the SigninOrReset client wrapper; ForgotPasswordForm moves to
its own file colocated with the new route. The new route opts back in
to indexing, matching the other public pages under the root-layout
noindex default.
